Move TrayIcon code to seprate file
