+++ UCR/Views/MainWindow.xaml.cs
@@ -22,15 +22,11 @@ namespace HidWizards.UCR.Views
 
     public partial class MainWindow : Window
     {
-        private Context Context { get; set; }
+        internal Context Context { get; set; }
+        private UCRTrayIcon TrayIcon;
         private readonly DashboardViewModel _dashboardViewModel;
         private CloseState WindowCloseState { get; set; }
         private Dictionary<Guid, ProfileWindow> ProfileWindows;
-        private System.Windows.Forms.NotifyIcon TrayIcon = new System.Windows.Forms.NotifyIcon();
-        private System.Windows.Forms.ToolStripItem ShowStrip = new System.Windows.Forms.ToolStripMenuItem();
-        private System.Windows.Forms.ToolStripItem StartLastProfileStrip = new System.Windows.Forms.ToolStripMenuItem();
-        private System.Windows.Forms.ToolStripItem StopProfileStrip = new System.Windows.Forms.ToolStripMenuItem();
-        private System.Windows.Forms.ToolStripItem ExitStrip = new System.Windows.Forms.ToolStripMenuItem();
 
         enum CloseState
         {
@@ -45,7 +41,7 @@ namespace HidWizards.UCR.Views
             DataContext = _dashboardViewModel;
             Context = context;
             ProfileWindows = new Dictionary<Guid, ProfileWindow>();
-            InitTrayIcon();
+            TrayIcon = new UCRTrayIcon(this);
             InitializeComponent();
         }
 
@@ -96,21 +92,10 @@ namespace HidWizards.UCR.Views
             }
         }
 
-        private void DeactivateProfile(object sender, RoutedEventArgs e)
+        internal void DeactivateProfile(object sender, RoutedEventArgs e)
         {
             if (Context.ActiveProfile == null) return;
             
-            if (!Context.SubscriptionsManager.DeactivateCurrentProfile())
-            {
-                // TODO Move to dialog
-                MessageBox.Show("The active Profile could not be deactivated, see the log for more details", "Profile failed to deactivate!", MessageBoxButton.OK, MessageBoxImage.Exclamation);
-            }
-        }
-
-        private void DeactivateProfile(object sender, EventArgs e)
-        {
-            if (Context.ActiveProfile == null) return;
-
             if (!Context.SubscriptionsManager.DeactivateCurrentProfile())
             {
                 // TODO Move to dialog
@@ -327,74 +312,7 @@ namespace HidWizards.UCR.Views
             var error = Marshal.GetLastWin32Error();
             MessageBox.Show($"Enabling message handling failed with the error: {error}");
         }
-
-        private void InitTrayIcon()
-        {
-            ShowStrip.Text = "Show";
-            ShowStrip.DisplayStyle = System.Windows.Forms.ToolStripItemDisplayStyle.Text;
-            ShowStrip.Click += ShowStrip_Click;
-
-            StartLastProfileStrip.Text = "Activate Last Profile";
-            StartLastProfileStrip.DisplayStyle = System.Windows.Forms.ToolStripItemDisplayStyle.Text;
-            StartLastProfileStrip.Click += StartLastProfileStrip_Click;
-            StartLastProfileStrip.Enabled = false;
-
-            StopProfileStrip.Text = "Stop Active Profile";
-            StopProfileStrip.DisplayStyle = System.Windows.Forms.ToolStripItemDisplayStyle.Text;
-            StopProfileStrip.Click += DeactivateProfile;
-            StopProfileStrip.Enabled = false;
-
-            ExitStrip.Text = "Exit";
-            ExitStrip.DisplayStyle = System.Windows.Forms.ToolStripItemDisplayStyle.Text;
-            ExitStrip.Click += ExitStrip_Click;
-
-            Context.ActiveProfileChangedEvent += Context_ActiveProfileChangedEvent;
-
-            TrayIcon.Text = "Universal Control Remapper";
-            TrayIcon.Icon = System.Drawing.Icon.ExtractAssociatedIcon(System.Windows.Forms.Application.ExecutablePath);
-            TrayIcon.DoubleClick += TrayIcon_OnDoubleClick;
-            TrayIcon.ContextMenuStrip = new System.Windows.Forms.ContextMenuStrip();
-            TrayIcon.ContextMenuStrip.ShowCheckMargin = false;
-            TrayIcon.ContextMenuStrip.ShowImageMargin = false;
-            TrayIcon.ContextMenuStrip.Items.AddRange(new System.Windows.Forms.ToolStripItem[] { ShowStrip, new System.Windows.Forms.ToolStripSeparator(), StartLastProfileStrip, StopProfileStrip, new System.Windows.Forms.ToolStripSeparator(), ExitStrip });
-            TrayIcon.Visible = true;
-        }
-
-        private void Context_ActiveProfileChangedEvent(Profile profile)
-        {
-            if(profile == null)
-            {
-                StopProfileStrip.Enabled = false;
-            }
-            else
-            {
-                StopProfileStrip.Enabled = true;
-                StartLastProfileStrip.Text = $"Activate Last Profile: {profile.Title}";
-                StartLastProfileStrip.Enabled = true;
-            }
-        }
-
-        private void ExitStrip_Click(object sender, EventArgs e)
-        {
-            Close();
-        }
-
-        private void StartLastProfileStrip_Click(object sender, EventArgs e)
-        {
-            Context.SubscriptionsManager.ActivateLastProfile();
-            StopProfileStrip.Enabled = true;
-        }
-
-        private void ShowStrip_Click(object sender, EventArgs e)
-        {
-            Show();
-            Activate();
-        }
-
-        private void TrayIcon_OnDoubleClick(object sender, EventArgs e)
-        {
-            Show();
-        }
+        
         private async void About_OnClick(object sender, RoutedEventArgs e)
         {
             var dialog = new AboutDialog();
